feat: updating the restriction for all admin apis

+++ src/modules/agent/agent.controller.ts
@@ -1,11 +1,16 @@
-import { api } from "encore.dev/api";
-import AgentService from "./agent.service";
+import { api, APIError } from "encore.dev/api";
+import { getAuthData } from "~encore/auth";
 import { MakeAgentDto, RemoveAgentDto, SuccessMessage, SuccessMessageWithAgent, SuccessMessageWithAgentAndTotalDiamond } from "./agent.interface";
+import AgentService from "./agent.service";
 
 // Make a user an agent
 export const makeAgent = api(
     { expose: true, auth: true, method: "POST", path: "/agents/make" },
     async (data: MakeAgentDto): Promise<SuccessMessage> => {
+        const role = getAuthData()!.role
+        if (role !== "ADMIN") {
+            throw APIError.permissionDenied("Only Admin is allowed")
+        }
         await AgentService.makeAgent(data.regNo);
         return { success: true, message: "User is now an agent" };
     }
@@ -15,6 +20,10 @@ export const makeAgent = api(
 export const removeAgent = api(
     { expose: true, auth: true, method: "POST", path: "/agents/remove" },
     async (data: RemoveAgentDto): Promise<SuccessMessage> => {
+        const role = getAuthData()!.role
+        if (role !== "ADMIN") {
+            throw APIError.permissionDenied("nly Admin is allowed")
+        }
         await AgentService.removeAgent(data.regNo);
         return { success: true, message: "User is no longer an agent" };
     }
